+++ ptit_dormitory_be/index.js
@@ -7,7 +7,8 @@ import { connectDB, sequelize } from './config/db.js';
 import './models/Assocication.js';
 import cors from 'cors';
 import errorHandler from './middleware/errorHandler.js';
-import swaggerDocs from './config/swagger.js';
+// import swaggerDocs from './config/swagger.js';
+import swaggerDocs from './swagger/index.js';
 
 import router from './routes/Routes.js';
 import fs from 'fs';
