+++ ptit_dormitory_be/swagger/auth.docs.js
@@ -1,0 +1,87 @@
+/**
+ * @swagger
+ * tags:
+ *   - name: Auth
+ *     description: Xác thực và đăng nhập hệ thống
+ */
+
+/**
+ * @swagger
+ * /api/auth/register:
+ *   post:
+ *     summary: Đăng ký người dùng mới
+ *     tags: [Auth]
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             required:
+ *               - email
+ *               - password
+ *               - first_name
+ *               - last_name
+ *             properties:
+ *               email:
+ *                 type: string
+ *                 example: user@example.com
+ *               password:
+ *                 type: string
+ *                 example: 123456
+ *               phone_number:
+ *                 type: string
+ *                 example: "0912345678"
+ *               dob:
+ *                 type: string
+ *                 format: date
+ *                 example: 2000-01-01
+ *               first_name:
+ *                 type: string
+ *                 example: Văn
+ *               last_name:
+ *                 type: string
+ *                 example: Nguyễn
+ *               status:
+ *                 type: integer
+ *                 example: 1
+ *               role_id:
+ *                 type: integer
+ *                 example: 1
+ *     responses:
+ *       201:
+ *         description: Đăng ký thành công
+ *       400:
+ *         description: Email đã tồn tại hoặc role_id không hợp lệ
+ */
+
+/**
+ * @swagger
+ * /api/auth/login:
+ *   post:
+ *     summary: Đăng nhập hệ thống
+ *     tags: [Auth]
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             required:
+ *               - email
+ *               - password
+ *             properties:
+ *               email:
+ *                 type: string
+ *                 example: user@example.com
+ *               password:
+ *                 type: string
+ *                 example: 123456
+ *     responses:
+ *       200:
+ *         description: Đăng nhập thành công, trả về token và thông tin người dùng
+ *       401:
+ *         description: Sai mật khẩu
+ *       404:
+ *         description: Không tìm thấy người dùng
+ */
